+++ src/main/java/com/smoky/bassshakertelemetry/client/accessibility/HudCueType.java
@@ -1,0 +1,10 @@
+package com.smoky.bassshakertelemetry.client.accessibility;
+
+public enum HudCueType {
+    DAMAGE,
+    EXPLOSION,
+    THUNDER,
+    BOSS,
+    WARDEN_HEARTBEAT,
+    LOW_HEALTH
+}